+++ extensions/1-click-confetti/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Confetti Changelog
 
+## [Custom emoji presets] - {PR_MERGE_DATE}
+
+- Added support for Raycast's custom confetti emojis.
+- Introduced configurable default icons and emoji presets
+
 ## [Add Confetti Command] - 2023-09-05
 
 Added Confetti command with sound to be used directly from Raycast and assigned to a hotkey.